♻️ refactor: 기존 application 안에 있던 dto 폴더를 adapter 하위로 변경
