+++ src/main/java/hello/pet/userservice/adapter/in/web/UserController.java
@@ -1,7 +1,7 @@
 package hello.pet.userservice.adapter.in.web;
 
-import hello.pet.userservice.application.dto.RegisterUserRequest;
-import hello.pet.userservice.application.dto.RegisterUserResponse;
+import hello.pet.userservice.adapter.in.web.dto.RegisterUserRequest;
+import hello.pet.userservice.adapter.in.web.dto.RegisterUserResponse;
 import hello.pet.userservice.application.port.in.CreateUserUseCase;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
