+++ src/main/java/hello/pet/userservice/application/port/in/CreateUserUseCase.java
@@ -1,7 +1,7 @@
 package hello.pet.userservice.application.port.in;
 
-import hello.pet.userservice.application.dto.RegisterUserRequest;
-import hello.pet.userservice.application.dto.RegisterUserResponse;
+import hello.pet.userservice.adapter.in.web.dto.RegisterUserRequest;
+import hello.pet.userservice.adapter.in.web.dto.RegisterUserResponse;
 import jakarta.validation.Valid;
 
 public interface CreateUserUseCase {
